gave user aproval rights to user 1

+++ hades/routes.py
@@ -107,20 +107,23 @@ def users():
 @app.route('/admin/users/<action>/<id>')
 @login_required
 def manage_user(action, id):
-    if action == 'approve':
-        user = User.query.filter_by(id=id).first()
-        user.approved = True
-        db.session.add(user)
-        db.session.commit()
-        flash(f'Approved user \'{user.username}\' successfully', 'success')
-    elif action == 'reject' or action == 'delete':
-        print('1')
-        username = User.query.filter_by(id=id).first().username
-        db.session.query(User).filter_by(id=id).delete()
-        db.session.commit()
-        if action == 'reject':
-            flash(f'Rejected user \'{username}\' successfully', 'success')
-        else:
+    if current_user.id == 1:
+        if action == 'approve':
+            user = User.query.filter_by(id=id).first()
+            user.approved = True
+            db.session.add(user)
+            db.session.commit()
+            flash(f'Approved user \'{user.username}\' successfully', 'success')
+        elif action == 'reject' or action == 'delete':
             print('1')
-            flash(f'Deleted user \'{username}\' successfully', 'success')
+            username = User.query.filter_by(id=id).first().username
+            db.session.query(User).filter_by(id=id).delete()
+            db.session.commit()
+            if action == 'reject':
+                flash(f'Rejected user \'{username}\' successfully', 'success')
+            else:
+                print('1')
+                flash(f'Deleted user \'{username}\' successfully', 'success')
+    else:
+        flash(f'You don\'t have sufficient rights.', 'danger')
     return redirect(url_for('users'))
